Fix basic test when build output isn't in ./target/debug
When your build output isn't stored in ./target/debug, the basic test
currently fails to find the binary it executes because of using a hardcoded
relative path. This can also cause problems if nvim-rs is part of a
workspace, as then the build output will be at the root of the workspace by
default.

I was hoping to just use `env!("CARGO_BIN_EXE_*")` for this, but it doesn't
seem there's any way to make use of this to find the examples directory.
Hopefully the next best thing is to just use the build output directory,
and walk upwards until we reach the examples directory.